Add: フラッシュメッセージを別のコンポーネントとして切り出した。
